+++ lib/ex_tdlib.ex
@@ -1,11 +1,11 @@
-defmodule TDLib do
+defmodule ExTDLib do
   @moduledoc """
   This module allow you to interact with and manage sessions.
   """
 
-  alias TDLib.Method
-  alias TDLib.Session
-  alias TDLib.Session.Registry
+  alias ExTDLib.Method
+  alias ExTDLib.Session
+  alias ExTDLib.Session.Registry
 
   @default_config %Method.SetTdlibParameters{
     database_encryption_key: nil,
@@ -30,7 +30,7 @@ defmodule TDLib do
   Configuration template for TDLib, to be modified and used as parameter of
   `open/3`.
 
-  See `TDLib.Object.TdlibParameters` and
+  See `ExTDLib.Object.TdlibParameters` and
   [core.telegram.org/tdlib/options](https://core.telegram.org/tdlib/options)
   for details. You can obtain an `:api_id` and an `:api_hash` on
   [my.telegram.org](https://my.telegram.org/) : they are required by TDLib,
@@ -70,13 +70,18 @@ defmodule TDLib do
   The parameter `msg` must be a struct (any map in reality) since it is
   directly encoded into JSON and transmitted via TDLib. You should use the
   structures generated from TDLib's documentation and provided by the
-  submodules of `TDLib.Object` and `TDLib.Methods`.
+  submodules of `ExTDLib.Object` and `ExTDLib.Methods`.
 
   Alternatively it is also possible to directly provide an already encoded
   binary or string, althrough you should not need it.
   """
   def transmit(session_name, msg) when is_map(msg) do
-    json = Poison.encode!(msg)
+    json =
+      msg
+      |> Map.delete(:__struct__)
+      |> Map.new(fn {k, v} -> {k, transform_struct(v)} end)
+      |> Jason.encode!()
+
     transmit(session_name, json)
   end
 
@@ -99,7 +104,7 @@ defmodule TDLib do
 
   @doc false
   def get_backend_binary do
-    config = Application.get_env(:tdlib, :backend_binary)
+    config = Application.get_env(:ex_tdlib, :backend_binary)
 
     case config do
       nil -> Path.join(Mix.Project.build_path(), "/lib/tdlib_json_cli/bin/tdlib_json_cli")
@@ -114,4 +119,12 @@ defmodule TDLib do
     first_letter = String.first(str)
     String.replace_prefix(str, first_letter, String.upcase(first_letter))
   end
+
+  defp transform_struct(map) when is_map(map) do
+    map
+    |> Map.delete(:__struct__)
+    |> Map.new(fn {k, v} -> {k, transform_struct(v)} end)
+  end
+
+  defp transform_struct(value), do: value
 end